Fix warnings

+++ SinusWave.h
@@ -2,13 +2,13 @@
 class SinusWave
 {
 public:
-	SinusWave(long long periodMs, float minValue, float maxValue);
-	void Tick(long long);
+	SinusWave(float periodMs, float minValue, float maxValue);
+	void Tick(float);
 	float GetValue();
 	bool IsExhausted();
 private:
-	long long m_periodLenghtMs;
-	long long m_currentTimePointMs;
+	float m_periodLenghtMs;
+	float m_currentTimePointMs;
 	float m_minValue;
 	float m_maxValue;
